Refactor registration tests
