+++ webapp/tests/views/test_logout.py
@@ -1,32 +1,15 @@
-from django.contrib.auth.hashers import make_password
 from django.contrib.sessions.models import Session
 from django.core.urlresolvers import reverse
 from django.test import TestCase
-from model_mommy import mommy
 
 from webapp.views import logout
-from webapp.models import User
-
-
-class UserHelperMixin:
-    def create_authenticated_user(self, password, **kwargs):
-        user = self.create_user(password, **kwargs)
-        logged_in = self.client.login(username=user.username,
-                                      password=password)
-        self.assertTrue(logged_in)
-        return user
-
-    def create_user(self, password, **kwargs):
-        return mommy.make(User,
-                          password=make_password(password),
-                          _fill_optional=['email'],
-                          **kwargs)
+from webapp.tests.user_helpers import UserHelperMixin
 
 
 class TestLogoutViewGET(UserHelperMixin, TestCase):
     def test_user_is_logged_out(self):
         password = '1234'
-        user = self.create_authenticated_user(password)
+        self.create_authenticated_user(password)
         #  just to be sure, we check a session exists after the user logs in
         self.assertTrue(Session.objects.all())
         response = self.client.get(reverse('logout'))
